requirements.txt
pyyaml
torch_utils
easydict
segmentation_models_pytorch
albumentations
tensorboardX
mmcv